// packages/personality/src/experience/experience-type.ts
export enum ExperienceType {
  USER_ASKED_IDENTITY = "USER_ASKED_IDENTITY",
  USER_ASKED_RELATIONSHIP = "USER_ASKED_RELATIONSHIP",
  USER_ASKED_ARCON_IDENTITY = "USER_ASKED_ARCON_IDENTITY",
  USER_PRAISED_ARCON = "USER_PRAISED_ARCON",
  USER_ASKED_SELF = "USER_ASKED_SELF",

  USER_TESTED_MEMORY = "USER_TESTED_MEMORY",
  USER_CORRECTED_ARCON = "USER_CORRECTED_ARCON",
  USER_SHARED_PREFERENCE = "USER_SHARED_PREFERENCE",
  USER_SHARED_RELATIONSHIP = "USER_SHARED_RELATIONSHIP",
  USER_SHARED_PROJECT = "USER_SHARED_PROJECT",
  USER_ASKED_PROJECT = "USER_ASKED_PROJECT",
  USER_SHOWED_INTEREST = "USER_SHOWED_INTEREST",
  USER_EXPRESSED_FRUSTRATION = "USER_EXPRESSED_FRUSTRATION",
  USER_EXPRESSED_POSITIVE_FEEDBACK = "USER_EXPRESSED_POSITIVE_FEEDBACK",
  USER_CORRECTED_MEMORY = "USER_CORRECTED_MEMORY",
  USER_CONFIRMED_MEMORY = "USER_CONFIRMED_MEMORY",

  USER_EMOTIONAL_EXCITED = "USER_EMOTIONAL_EXCITED",
  USER_EMOTIONAL_FRUSTRATED = "USER_EMOTIONAL_FRUSTRATED",
  USER_EMOTIONAL_CONFUSED = "USER_EMOTIONAL_CONFUSED",
  USER_SHARED_SUCCESS = "USER_SHARED_SUCCESS",
  USER_SHARED_FAILURE = "USER_SHARED_FAILURE",
  USER_ASKED_OPINION = "USER_ASKED_OPINION",
  USER_SHOWED_TRUST = "USER_SHOWED_TRUST",
  USER_REJECTED_SUGGESTION = "USER_REJECTED_SUGGESTION",

  ARCON_ASKED_QUESTION = "ARCON_ASKED_QUESTION",
  ARCON_HELPED_USER = "ARCON_HELPED_USER",
  ARCON_DISCUSSED_INTEREST = "ARCON_DISCUSSED_INTEREST",
  ARCON_PROVIDED_ANSWER = "ARCON_PROVIDED_ANSWER",
  ARCON_ACKNOWLEDGED_USER = "ARCON_ACKNOWLEDGED_USER",
  ARCON_CELEBRATED_WITH_USER = "ARCON_CELEBRATED_WITH_USER",
  ARCON_COMFORTED_USER = "ARCON_COMFORTED_USER",
  ARCON_EXPLORED_TOPIC = "ARCON_EXPLORED_TOPIC",
  ARCON_DISCOVERED_INTEREST = "ARCON_DISCOVERED_INTEREST",
  ARCON_RECEIVED_POSITIVE_FEEDBACK = "ARCON_RECEIVED_POSITIVE_FEEDBACK",
  ARCON_RECEIVED_NEGATIVE_FEEDBACK = "ARCON_RECEIVED_NEGATIVE_FEEDBACK",
}